feat: app tools command
