+++ packages/solutions/app-tools/src/index.ts
@@ -3,116 +3,39 @@ import { lintPlugin } from '@modern-js/plugin-lint';
 import {
   cleanRequireCache,
   emptyDir,
-  Command,
   getCommand,
   getArgv,
   fs,
   NESTED_ROUTE_SPEC_FILE,
-  newAction,
-  upgradeAction,
 } from '@modern-js/utils';
-import { castArray } from '@modern-js/utils/lodash';
-import { CliPlugin, PluginAPI } from '@modern-js/core';
+import { CliPlugin } from '@modern-js/core';
 import { getLocaleLanguage } from '@modern-js/plugin-i18n/language-detector';
 import analyzePlugin from './analyze';
 import initializePlugin from './initialize';
 import { AppTools } from './types';
 import { hooks } from './hooks';
-import { i18n, localeKeys } from './locale';
+import { i18n } from './locale';
 import serverBuildPlugin from './plugins/serverBuild';
 import deployPlugin from './plugins/deploy';
-import type {
-  DevOptions,
-  BuildOptions,
-  DeployOptions,
-  InspectOptions,
-} from './utils/types';
 import { restart } from './utils/restart';
 import { generateWatchFiles } from './utils/generateWatchFiles';
+import {
+  buildCommand,
+  deployCommand,
+  devCommand,
+  inspectCommand,
+  newCommand,
+  serverCommand,
+  upgradeCommand,
+} from './commands';
 
 export { mergeConfig } from '@modern-js/core';
-export { dev } from './commands';
 export * from './defineConfig';
 // eslint-disable-next-line import/export
 export * from './types';
 
 // eslint-disable-next-line import/export
 export type { RuntimeUserConfig } from './types/config';
-
-export const devCommand = async (
-  program: Command,
-  api: PluginAPI<AppTools<'shared'>>,
-) => {
-  const runner = api.useHookRunners();
-  const devToolMetas = await runner.registerDev();
-
-  const devProgram = program
-    .command('dev')
-    .alias('start')
-    .usage('[options]')
-    .description(i18n.t(localeKeys.command.dev.describe))
-    .option('-c --config <config>', i18n.t(localeKeys.command.shared.config))
-    .option('-e --entry [entry...]', i18n.t(localeKeys.command.dev.entry))
-    .option('--analyze', i18n.t(localeKeys.command.shared.analyze))
-    .option('--api-only', i18n.t(localeKeys.command.dev.apiOnly))
-    .option('--web-only', i18n.t(localeKeys.command.dev.webOnly))
-    .action(async (options: DevOptions) => {
-      const { dev } = await import('./commands/dev');
-      await dev(api, options);
-    });
-
-  for (const meta of devToolMetas) {
-    if (!meta.subCommands) {
-      continue;
-    }
-
-    for (const subCmd of meta.subCommands) {
-      devProgram.command(subCmd).action(async (options: DevOptions = {}) => {
-        const { appDirectory } = api.useAppContext();
-        const { isTypescript } = await import('@modern-js/utils');
-
-        await runner.beforeDevTask(meta);
-        await meta.action(options, {
-          isTsProject: isTypescript(appDirectory),
-        });
-      });
-    }
-  }
-};
-
-export const buildCommand = async (
-  program: Command,
-  api: PluginAPI<AppTools<'shared'>>,
-) => {
-  const runner = api.useHookRunners();
-  const platformBuilders = await runner.registerBuildPlatform();
-
-  const buildProgram = program
-    .command('build')
-    .usage('[options]')
-    .description(i18n.t(localeKeys.command.build.describe))
-    .option('-c --config <config>', i18n.t(localeKeys.command.shared.config))
-    .option('--analyze', i18n.t(localeKeys.command.shared.analyze))
-    .action(async (options: BuildOptions) => {
-      const { build } = await import('./commands/build');
-      await build(api, options);
-    });
-
-  for (const platformBuilder of platformBuilders) {
-    const platforms = castArray(platformBuilder.platform);
-    for (const platform of platforms) {
-      buildProgram.command(platform).action(async () => {
-        const { appDirectory } = api.useAppContext();
-        const { isTypescript } = await import('@modern-js/utils');
-
-        await runner.beforeBuildPlatform(platformBuilders);
-        await platformBuilder.build(platform, {
-          isTsProject: isTypescript(appDirectory),
-        });
-      });
-    }
-  }
-};
 
 export type AppToolsOptions = {
   /**
@@ -185,107 +108,12 @@ export const appTools = (
       },
       async commands({ program }) {
         await devCommand(program, api);
-
         await buildCommand(program, api);
-
-        program
-          .command('serve')
-          .usage('[options]')
-          .description(i18n.t(localeKeys.command.serve.describe))
-          .option('--api-only', i18n.t(localeKeys.command.dev.apiOnly))
-          .option(
-            '-c --config <config>',
-            i18n.t(localeKeys.command.shared.config),
-          )
-          .action(async () => {
-            const { start } = await import('./commands/serve');
-            await start(api);
-          });
-
-        program
-          .command('deploy')
-          .usage('[options]')
-          .option(
-            '-c --config <config>',
-            i18n.t(localeKeys.command.shared.config),
-          )
-          .option(
-            '-s --skip-build',
-            i18n.t(localeKeys.command.shared.skipBuild),
-          )
-          .description(i18n.t(localeKeys.command.deploy.describe))
-          .action(async (options: DeployOptions) => {
-            if (!options.skipBuild) {
-              const { build } = await import('./commands/build');
-              await build(api);
-            }
-
-            const { deploy } = await import('./commands/deploy');
-            await deploy(api, options);
-            // eslint-disable-next-line no-process-exit
-            process.exit(0);
-          });
-
-        program
-          .command('new')
-          .usage('[options]')
-          .description(i18n.t(localeKeys.command.new.describe))
-          .option(
-            '--config-file <configFile>',
-            i18n.t(localeKeys.command.shared.config),
-          )
-          .option('--lang <lang>', i18n.t(localeKeys.command.new.lang))
-          .option(
-            '-c, --config <config>',
-            i18n.t(localeKeys.command.new.config),
-          )
-          .option('-d, --debug', i18n.t(localeKeys.command.new.debug), false)
-          .option('--dist-tag <tag>', i18n.t(localeKeys.command.new.distTag))
-          .option('--registry', i18n.t(localeKeys.command.new.registry))
-          .option(
-            '--no-need-install',
-            i18n.t(localeKeys.command.shared.noNeedInstall),
-          )
-          .action(async (options: any) => {
-            await newAction(
-              {
-                ...options,
-                locale: options.lang || locale,
-              },
-              'mwa',
-            );
-          });
-
-        program
-          .command('inspect')
-          .description('inspect the internal configs')
-          .option(
-            `--env <env>`,
-            i18n.t(localeKeys.command.inspect.env),
-            'development',
-          )
-          .option(
-            '--output <output>',
-            i18n.t(localeKeys.command.inspect.output),
-            '/',
-          )
-          .option('--verbose', i18n.t(localeKeys.command.inspect.verbose))
-          .option(
-            '-c --config <config>',
-            i18n.t(localeKeys.command.shared.config),
-          )
-          .action(async (options: InspectOptions) => {
-            const { inspect } = await import('./commands/inspect');
-            inspect(api, options);
-          });
-
-        program
-          .command('upgrade')
-          .allowUnknownOption()
-          .option('-h --help', 'Show help') // In order to upgrade help work.
-          .action(async () => {
-            await upgradeAction();
-          });
+        serverCommand(program, api);
+        deployCommand(program, api);
+        newCommand(program, locale);
+        inspectCommand(program, api);
+        upgradeCommand(program);
       },
 
       async prepare() {
